+++ data/command.txt
@@ -1,0 +1,1 @@
+D | 1 | return book | 2/12/2019 1800
