+++ src/main/java/Deadline.java
@@ -1,30 +1,33 @@
 package main.java;
 
-import java.util.concurrent.DelayQueue;
+import java.time.LocalDateTime;
 
 /**
  * Deadline is a type of Task that contains addition.
  */
 
 public class Deadline extends Task {
-  String type = "D";
-  String info;
+  private String type = "D";
+  private String info;
+  private LocalDateTime dateInfo;
+  
 
   /**
    * Constructor for deadline.
    * 
    * @param name name of task.
-   * @param info additional info of the task.
+   * @param dateInfo additional dateInfo of the task.
    */
 
-  Deadline(String name, String info) {
+  Deadline(String name, String dateInfo) {
     super(name);
-    this.info = info;
+    info = dateInfo;
+    this.dateInfo = LocalDateTime.parse(dateInfo, this.dateTimeFormatIn);
   }
 
-  Deadline(String name, boolean marked, String info) {
+  Deadline(String name, boolean marked, String dateInfo) {
     super(name);
-    this.info = info;
+    this.dateInfo = LocalDateTime.parse(dateInfo, this.dateTimeFormatIn);
     this.isMarked = marked;
   }
 
@@ -39,6 +42,6 @@ public class Deadline extends Task {
 
   @Override
   String display() {
-    return "[" + this.type + "] " + "[" + markDisplay() + "] " + this.name + " (by " + info + ")";
+    return "[" + this.type + "] " + "[" + markDisplay() + "] " + this.name + " (by " + dateInfo.format(this.dateTimeFormatterOut) + ")";
   }
 }
